+++ src/main/java/com/wigell/wigellcarrental/services/CustomerService.java
@@ -1,7 +1,9 @@
 package com.wigell.wigellcarrental.services;
 
 import com.wigell.wigellcarrental.entities.Customer;
+import com.wigell.wigellcarrental.entities.Order;
 
+import java.security.Principal;
 import java.util.List;
 
 //SA
@@ -10,4 +12,6 @@ public interface CustomerService {
 
     // WIG-27-SJ
     Customer getCustomerById(Long id);
+
+    List<Order> getOrders(Principal principal); //WIG-22-AA
 }
