+++ src/main/resources/data.sql
@@ -23,4 +23,4 @@ INSERT INTO orders (booked_at, start_date, end_date, car_id, customer_id, total_
 ('2025-01-15', '2025-01-22', '2025-01-26', 2, 2, 2796.00, false),
 ('2025-04-01', '2025-04-05', '2025-04-15', 3, 3, 8990.00, true),
 ('2025-03-28', '2025-04-01', '2025-04-07', 4, 4, 4494.00, true),
-('2025-04-05', '2025-04-06', '2025-04-12', 5, 5, 3294.00, true);
+('2025-04-05', '2025-06-06', '2025-06-12', 5, 5, 3294.00, true);
